debugged Please enter the commit message for your changes. Lines starting

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -11,7 +11,7 @@
 
 size_t print_list(const list_t *h)
 {
-	size_t nodes = 0;
+	unsigned int nodes = 0;
 
 	if (h == NULL)
 		return (0);
@@ -24,7 +24,9 @@ size_t print_list(const list_t *h)
 		{
 			printf("[%u] %s\n", h->len, h->str);
 		}
+
 		nodes++;
+
 		h = h->next;
 	}
